OTT-1160: move NewMetricsRegistry in OW specific files
